requirements.txt
Flask~=2.2.2
qrcode[pil]~=7.3.1
django-qrcode~=0.3
requests~=2.28.2
mysql-connector-python~=8.0.32
beautifulsoup4~=4.11.1
setuptools
pycoingecko~=3.1.0
btcpay-python
lxml
Werkzeug~=2.2.2